# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WebForge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List my sites
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create a site (queues analysis)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get site by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete a site
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /sites/{id}/confirm:
    post:
      operationId: confirmSite
      tags: [sites]
      summary: Confirm the analysis/plan and start the build
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}/edit:
    post:
      operationId: editSite
      tags: [sites]
      summary: Edit site with a new prompt
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditSiteInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}/retry:
    post:
      operationId: retrySite
      tags: [sites]
      summary: Retry the last failed step
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}/messages:
    get:
      operationId: listSiteMessages
      tags: [messages]
      summary: List the chat between the user and the agent for a site
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendSiteMessage
      tags: [messages]
      summary: Send a message in the agent chat for a site
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
  /sites/{id}/domain:
    post:
      operationId: setSiteDomain
      tags: [sites]
      summary: Attach a custom domain to a site
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDomainInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: removeSiteDomain
      tags: [sites]
      summary: Detach the custom domain from a site
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}/domain/verify:
    post:
      operationId: verifySiteDomain
      tags: [sites]
      summary: Verify the custom domain by checking DNS records
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    SiteAnalysis:
      type: object
      properties:
        type:
          type: string
          enum: [website, bot, backend, tool]
        intent: { type: string }
        audience: { type: string, nullable: true }
        features:
          type: array
          items: { type: string }
        pages:
          type: array
          items: { type: string }
        styleHints:
          type: array
          items: { type: string }
      required: [type, intent, features, pages, styleHints]
    SitePlanPage:
      type: object
      properties:
        path: { type: string }
        title: { type: string }
        purpose: { type: string }
        sections:
          type: array
          items: { type: string }
      required: [path, title, purpose, sections]
    SitePlan:
      type: object
      properties:
        type:
          type: string
          enum: [website, bot, backend, tool]
        summary: { type: string }
        pages:
          type: array
          items: { $ref: "#/components/schemas/SitePlanPage" }
        styles:
          type: object
          properties:
            palette: { type: string }
            mood: { type: string }
          required: [palette, mood]
        features:
          type: array
          items: { type: string }
        notes:
          type: array
          items: { type: string }
      required: [type, summary, pages, styles, features, notes]
    Site:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        prompt: { type: string }
        status:
          type: string
          enum:
            [
              queued,
              analyzing,
              awaiting_confirmation,
              building,
              ready,
              failed,
            ]
        progress: { type: integer, minimum: 0, maximum: 100 }
        message: { type: string, nullable: true }
        error: { type: string, nullable: true }
        previewUrl: { type: string, nullable: true }
        publicUrl: { type: string, nullable: true }
        coverColor: { type: string, nullable: true }
        files:
          type: array
          items: { type: string }
        analysis:
          allOf:
            - $ref: "#/components/schemas/SiteAnalysis"
          nullable: true
        plan:
          allOf:
            - $ref: "#/components/schemas/SitePlan"
          nullable: true
        customDomain: { type: string, nullable: true }
        customDomainStatus:
          type: string
          enum: [pending, verified, failed]
          nullable: true
        customDomainError: { type: string, nullable: true }
        customDomainTxtName: { type: string, nullable: true }
        customDomainTxtValue: { type: string, nullable: true }
        customDomainTarget: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        [
          id,
          name,
          slug,
          prompt,
          status,
          progress,
          files,
          createdAt,
          updatedAt,
        ]
    CreateSiteInput:
      type: object
      properties:
        prompt:
          type: string
          minLength: 4
          maxLength: 1000
        name:
          type: string
          maxLength: 80
          nullable: true
      required: [prompt]
    EditSiteInput:
      type: object
      properties:
        prompt:
          type: string
          minLength: 4
          maxLength: 1000
      required: [prompt]
    SetDomainInput:
      type: object
      properties:
        domain:
          type: string
          minLength: 3
          maxLength: 253
      required: [domain]
    SendMessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
      required: [content]
    Message:
      type: object
      properties:
        id: { type: string }
        siteId: { type: string }
        role:
          type: string
          enum: [user, agent, system]
        kind:
          type: string
          enum:
            [
              text,
              analysis,
              plan,
              awaiting_confirmation,
              log,
              build_started,
              build_progress,
              build_done,
              build_failed,
            ]
        content: { type: string }
        data:
          type: object
          additionalProperties: true
          nullable: true
        createdAt: { type: string, format: date-time }
      required: [id, siteId, role, kind, content, createdAt]
...