# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WebForge API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageData } from "./messageData";
import type { MessageKind } from "./messageKind";
import type { MessageRole } from "./messageRole";

export interface Message {
  id: string;
  siteId: string;
  role: MessageRole;
  kind: MessageKind;
  content: string;
  data?: MessageData;
  createdAt: Date;
}