# lib/api-zod/src/generated/types/sendMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WebForge API
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageInput {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
}